use go generate to generate RegionCodeInfo
